+++ cmd/suggestions.go
@@ -131,3 +131,18 @@ func suggestUsers(_ *cobra.Command, _ []string, toComplete string, c *api.Client
 
 	return va, nil
 }
+
+func suggestDescription(_ *cobra.Command, _ []string, toComplete string, c *api.Client) (completion.ValidArgs, error) {
+	if viper.GetBool(DESCR_AUTOCOMP) {
+		return completion.EmptyValidArgs(), nil
+	}
+
+	dc := newDescriptionCompleter(
+		c,
+		viper.GetString(WORKSPACE),
+		viper.GetString(USER_ID),
+		viper.GetInt(DESCR_AUTOCOMP_DAYS),
+	)
+
+	return completion.ValigsArgsSlide(dc.suggestFn(toComplete)), nil
+}